Move ingress from ingress to template

+++ scripts/test-definitions.sh
@@ -5,21 +5,23 @@ cd modules/htc-org/manifests/
 templates=$(ls .)
 for template in $templates;
 do
-  echo "## Testing ${template}:"
-  cp ../../../scripts/test.yaml $template/test.yaml
-  cd $template
-  
-  template=$template yq -i '.entity.type = env(template)' test.yaml
-  
-  yq -i '.entity.driver_inputs.values = load("definition-values.yaml")' test.yaml
-  
-  tests=$(ls tests/*-inputs.yaml)
-  for test in $tests;
-  do
-    echo "## ${template} tested with ${test}."
-    humctl resources test-definition test.yaml --inputs $test
-  done
+  if [ "${template}" != "k8s-ingress" ]; then
+    echo "## Testing ${template}:"
+    cp ../../../scripts/test.yaml $template/test.yaml
+    cd $template
+    
+    template=$template yq -i '.entity.type = env(template)' test.yaml
+    
+    yq -i '.entity.driver_inputs.values = load("definition-values.yaml")' test.yaml
+    
+    tests=$(ls tests/*-inputs.yaml)
+    for test in $tests;
+    do
+      echo "## ${template} tested with ${test}."
+      humctl resources test-definition test.yaml --inputs $test
+    done
 
-  rm test.yaml
-  cd ..
+    rm test.yaml
+    cd ..
+  fi
 done
